backend/requirements.txt
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
mysql-connector-python>=8.0.0,<9.0
pyjwt>=2.8.0,<3.0
pydantic-settings>=2.1.0,<3.0
python-multipart>=0.0.6,<1.0
cachetools>=5.3.0,<6.0

# dev / test
pytest>=8.0.0,<9.0
httpx>=0.27.0,<1.0
